requirements.txt
Django==5.2.7
whitenoise==6.6.0
tensorflow-cpu==2.16.1
Pillow==10.4.0
numpy==1.26.4
